pulled ops out of pc_register()
