+++ suite2p/registration/__init__.py
@@ -1,3 +1,1 @@
-from .pc import get_pc_metrics
-from .main import register_binary
-from .zalign import compute_zpos
+from .main import register_binary, get_pc_metrics, compute_zpos
